+++ madeasy/parser/__init__.py
@@ -7,4 +7,4 @@ from textx.metamodel import metamodel_from_file
 PARSER_APP = os.path.dirname(os.path.abspath(__file__))
 parser_file = os.path.join(PARSER_APP, 'parser.tx')
 
-parser_mm = metamodel_from_file(parser_file, debug=False)
+parser_mm = metamodel_from_file(parser_file, debug=False, ignore_case=True)